+++ sdk/resourcemanager/costmanagement/armcostmanagement/generatedetailedcostreportoperationresults_client.go
@@ -20,7 +20,7 @@ import (
 // group.
 // Don't use this type directly, use NewGenerateDetailedCostReportOperationResultsClient() instead.
 //
-// Generated from API version 2025-03-01
+// Generated from API version 2026-06-01
 type GenerateDetailedCostReportOperationResultsClient struct {
 	internal *arm.Client
 }
@@ -100,7 +100,7 @@ func (client *GenerateDetailedCostReportOperationResultsClient) getCreateRequest
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", version20250301)
+	reqQP.Set("api-version", version20260601)
 	req.Raw().URL.RawQuery = strings.ReplaceAll(reqQP.Encode(), "+", "%20")
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil