+++ sdk/resourcemanager/costmanagement/armcostmanagement/generatedetailedcostreportoperationstatus_client_example_test.go
@@ -11,7 +11,7 @@ import (
 	"log"
 )
 
-// Generated from example definition: 2025-03-01/GenerateDetailedCostReportOperationStatusBySubscriptionScope.json
+// Generated from example definition: 2026-06-01/GenerateDetailedCostReportOperationStatusBySubscriptionScope.json
 func ExampleGenerateDetailedCostReportOperationStatusClient_Get() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {